+++ saleor/graphql/core/mutations.py
@@ -21,6 +21,7 @@ from ..account.types import User
 from ..utils import get_nodes
 from .types import Error, MetaInput, MetaPath, Upload
 from .utils import from_global_id_strict_type, snake_to_camel_case
+from .utils.error_codes import get_error_code_from_error
 
 registry = get_global_registry()
 
@@ -42,21 +43,38 @@ def get_output_fields(model, return_field_name):
     return fields
 
 
+def get_error_fields(error_type_class, error_type_field):
+    """ """
+    return {
+        error_type_field: graphene.Field(
+            graphene.List(
+                graphene.NonNull(error_type_class),
+                description="List of errors that occurred executing the mutation.",
+            )
+        )
+    }
+
+
 def validation_error_to_error_type(validation_error: ValidationError) -> list:
     """Convert a ValidationError into a list of Error types."""
     err_list = []
     if hasattr(validation_error, "error_dict"):
         # convert field errors
-        for field, field_errors in validation_error.message_dict.items():
+        for field, field_errors in validation_error.error_dict.items():
+            field = None if field == NON_FIELD_ERRORS else snake_to_camel_case(field)
             for err in field_errors:
-                field = (
-                    None if field == NON_FIELD_ERRORS else snake_to_camel_case(field)
+                err_list.append(
+                    (
+                        Error(field=field, message=err.messages[0]),
+                        get_error_code_from_error(err),
+                    )
                 )
-                err_list.append(Error(field=field, message=err))
     else:
         # convert non-field errors
         for err in validation_error.error_list:
-            err_list.append(Error(message=err.message))
+            err_list.append(
+                (Error(message=err.messages[0]), get_error_code_from_error(err))
+            )
     return err_list
 
 
@@ -77,9 +95,14 @@ class BaseMutation(graphene.Mutation):
 
     @classmethod
     def __init_subclass_with_meta__(
-        cls, description=None, permissions: Tuple = None, _meta=None, **options
+        cls,
+        description=None,
+        permissions: Tuple = None,
+        _meta=None,
+        error_type_class=None,
+        error_type_field=None,
+        **options,
     ):
-
         if not _meta:
             _meta = MutationOptions(cls)
 
@@ -95,9 +118,15 @@ class BaseMutation(graphene.Mutation):
             )
 
         _meta.permissions = permissions
+        _meta.error_type_class = error_type_class
+        _meta.error_type_field = error_type_field
         super().__init_subclass_with_meta__(
             description=description, _meta=_meta, **options
         )
+        if error_type_class and error_type_field:
+            cls._meta.fields.update(
+                get_error_fields(error_type_class, error_type_field)
+            )
 
     @classmethod
     def _update_mutation_arguments_and_fields(cls, arguments, fields):
@@ -136,11 +165,17 @@ class BaseMutation(graphene.Mutation):
 
             node = cls.get_node_by_pk(info, graphene_type=only_type, pk=pk, qs=qs)
         except (AssertionError, GraphQLError) as e:
-            raise ValidationError({field: str(e)})
+            raise ValidationError(
+                {field: ValidationError(str(e), code="graphql_error")}
+            )
         else:
             if node is None:
                 raise ValidationError(
-                    {field: "Couldn't resolve to a node: %s" % node_id}
+                    {
+                        field: ValidationError(
+                            "Couldn't resolve to a node: %s" % node_id, code="not_found"
+                        )
+                    }
                 )
         return node
 
@@ -149,7 +184,9 @@ class BaseMutation(graphene.Mutation):
         try:
             instances = get_nodes(ids, only_type, qs=qs)
         except GraphQLError as e:
-            raise ValidationError({field: str(e)})
+            raise ValidationError(
+                {field: ValidationError(str(e), code="graphql_error")}
+            )
         return instances
 
     @classmethod
@@ -234,12 +271,30 @@ class BaseMutation(graphene.Mutation):
                 response.errors = []
             return response
         except ValidationError as e:
-            errors = validation_error_to_error_type(e)
-            return cls(errors=errors)
+            return cls.handle_errors(e)
 
     @classmethod
     def perform_mutation(cls, root, info, **data):
         pass
+
+    @classmethod
+    def handle_errors(cls, error: ValidationError, **extra):
+        errors = validation_error_to_error_type(error)
+        return cls.handle_typed_errors(errors, **extra)
+
+    @classmethod
+    def handle_typed_errors(cls, errors: list, **extra):
+        """Return class instance with errors."""
+        if (
+            cls._meta.error_type_class is not None
+            and cls._meta.error_type_field is not None
+        ):
+            typed_errors = [
+                cls._meta.error_type_class(field=e.field, message=e.message, code=code)
+                for e, code in errors
+            ]
+            extra.update({cls._meta.error_type_field: typed_errors})
+        return cls(errors=[e[0] for e in errors], **extra)
 
 
 class ModelMutation(BaseMutation):
@@ -496,7 +551,8 @@ class BaseBulkMutation(BaseMutation):
 
         count, errors = cls.perform_mutation(root, info, **data)
         if errors:
-            errors = validation_error_to_error_type(errors)
+            return cls.handle_errors(errors, count=count)
+
         return cls(errors=errors, count=count)