+++ saleor/graphql/payment/mutations.py
@@ -5,6 +5,7 @@ from django.core.exceptions import ValidationError
 from ...core.taxes import zero_taxed_money
 from ...core.utils import get_client_ip
 from ...payment import PaymentError, models
+from ...payment.error_codes import PaymentErrorCode
 from ...payment.utils import (
     create_payment,
     gateway_capture,
@@ -17,6 +18,7 @@ from ..account.types import AddressInput
 from ..checkout.types import Checkout
 from ..core.mutations import BaseMutation
 from ..core.scalars import Decimal
+from ..core.types import common as common_types
 from ..core.utils import from_global_id_strict_type
 from .enums import PaymentGatewayEnum
 from .types import Payment
@@ -61,6 +63,8 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
 
     class Meta:
         description = "Create a new payment for given checkout."
+        error_type_class = common_types.PaymentError
+        error_type_field = "payment_errors"
 
     @classmethod
     def perform_mutation(cls, _root, info, checkout_id, **data):
@@ -77,7 +81,12 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
             billing_address = cls.validate_address(data["billing_address"])
         if billing_address is None:
             raise ValidationError(
-                {"billing_address": "No billing address associated with this checkout."}
+                {
+                    "billing_address": ValidationError(
+                        "No billing address associated with this checkout.",
+                        code=PaymentErrorCode.BILLING_ADDRESS_NOT_SET,
+                    )
+                }
             )
 
         checkout_total = (
@@ -91,8 +100,11 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
         if amount < checkout_total.gross.amount:
             raise ValidationError(
                 {
-                    "amount": "Partial payments are not allowed, amount should be "
-                    "equal checkout's total."
+                    "amount": ValidationError(
+                        "Partial payments are not allowed, amount should be "
+                        "equal checkout's total.",
+                        code=PaymentErrorCode.PARTIAL_PAYMENT_NOT_ALLOWED,
+                    )
                 }
             )
 
@@ -122,6 +134,8 @@ class PaymentCapture(BaseMutation):
     class Meta:
         description = "Captures the authorized payment amount"
         permissions = ("order.manage_orders",)
+        error_type_class = common_types.PaymentError
+        error_type_field = "payment_errors"
 
     @classmethod
     def perform_mutation(cls, _root, info, payment_id, amount=None):
@@ -131,7 +145,7 @@ class PaymentCapture(BaseMutation):
         try:
             gateway_capture(payment, amount)
         except PaymentError as e:
-            raise ValidationError(str(e))
+            raise ValidationError(str(e), code=PaymentErrorCode.PAYMENT_ERROR)
         return PaymentCapture(payment=payment)
 
 
@@ -139,6 +153,8 @@ class PaymentRefund(PaymentCapture):
     class Meta:
         description = "Refunds the captured payment amount"
         permissions = ("order.manage_orders",)
+        error_type_class = common_types.PaymentError
+        error_type_field = "payment_errors"
 
     @classmethod
     def perform_mutation(cls, _root, info, payment_id, amount=None):
@@ -148,7 +164,7 @@ class PaymentRefund(PaymentCapture):
         try:
             gateway_refund(payment, amount=amount)
         except PaymentError as e:
-            raise ValidationError(str(e))
+            raise ValidationError(str(e), code=PaymentErrorCode.PAYMENT_ERROR)
         return PaymentRefund(payment=payment)
 
 
@@ -161,6 +177,8 @@ class PaymentVoid(BaseMutation):
     class Meta:
         description = "Voids the authorized payment"
         permissions = ("order.manage_orders",)
+        error_type_class = common_types.PaymentError
+        error_type_field = "payment_errors"
 
     @classmethod
     def perform_mutation(cls, _root, info, payment_id):
@@ -170,7 +188,7 @@ class PaymentVoid(BaseMutation):
         try:
             gateway_void(payment)
         except PaymentError as e:
-            raise ValidationError(str(e))
+            raise ValidationError(str(e), code=PaymentErrorCode.PAYMENT_ERROR)
         return PaymentVoid(payment=payment)
 
 
@@ -182,6 +200,8 @@ class PaymentSecureConfirm(BaseMutation):
 
     class Meta:
         description = "Confirms payment in two step process like 3D secure"
+        error_type_class = common_types.PaymentError
+        error_type_field = "payment_errors"
 
     @classmethod
     def perform_mutation(cls, _root, info, payment_id):
@@ -191,5 +211,5 @@ class PaymentSecureConfirm(BaseMutation):
         try:
             gateway_confirm(payment)
         except PaymentError as e:
-            raise ValidationError(str(e))
+            raise ValidationError(str(e), code=PaymentErrorCode.PAYMENT_ERROR)
         return PaymentSecureConfirm(payment=payment)